Fix "Expected Lua number" error.
I had to replace 'self.winId' with 0 to fix the game. As far as I see 'self.winId' is 
not used and is always nil. Maybe in previous versions of nvim it was fine to 
pass nil into nvim_win_set_cursor but in NVIM v0.8.0 it fails with the error 
"game-runner.lua:306: Expected Lua number" and the game stops. That is it 
doesn't get to the next round.